# feat(event pool): changed realisation to sync.Pool based