# Description
Changed realisation to sync.Pool based

**Pros**: increases performance
**Cons**: unavailable to get dump of all events bodies

# Benchmarks
## Old
```
goos: darwin
goarch: arm64
pkg: github.com/ozontech/file.d/pipeline
BenchmarkEventPoolOneGoroutine-8        58907269                20.13 ns/op            0 B/op          0 allocs/op
BenchmarkEventPoolManyGoroutines-8           138           7510056 ns/op            3509 B/op         43 allocs/op
BenchmarkEventPoolSlowestPath-8              164           7521373 ns/op           26861 B/op       1031 allocs/op
PASS
ok      github.com/ozontech/file.d/pipeline     13.210s
```
## New
```
goos: darwin
goarch: arm64
pkg: github.com/ozontech/file.d/pipeline
BenchmarkEventPoolOneGoroutine-8        43880164                26.87 ns/op            0 B/op          0 allocs/op
BenchmarkEventPoolManyGoroutines-8           873           1374570 ns/op             732 B/op         14 allocs/op
BenchmarkEventPoolSlowestPath-8             1647            898335 ns/op           24509 B/op       1005 allocs/op
PASS
ok      github.com/ozontech/file.d/pipeline     12.118s
```